+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 This project uses [Semantic Versioning 2.0.0](http://semver.org/), the format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## 0.3.2 - 2026-01-29
+
+### Changed
+
+- Update development dependencies.
+
 ## 0.3.1 - 2026-01-29
 
 ### Added